+++ spamcontrol/domainprovider.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"fmt"
 	"net/http"
+	"sort"
 	"strings"
 )
 
@@ -14,15 +15,52 @@ type SpamDomainProvider interface {
 	GetSpamDomains() ([]string, error)
 }
 
-// The staticSpamDomains provides a static list
+// NewAggregateProvider returns a new aggregate providers for the given list of SpamDomainProviders.
+func NewAggregateProvider(providers []SpamDomainProvider) SpamDomainProvider {
+	return AggregateProvider(providers)
+}
+
+// NewRemoteSpamDomainProvider creates a new remote spam domain provider for the given URL.
+func NewRemoteSpamDomainProvider(url string) SpamDomainProvider {
+	return &remoteSpamDomains{url}
+}
+
+// NewLocalSpamDomainProvider creates a new remote spam domain provider for the file path.
+func NewLocalSpamDomainProvider(filePath string) SpamDomainProvider {
+	return &localSpamDomains{filePath}
+}
+
+// An AggregateProvider combines multiple SpamDomainProviders.
+type AggregateProvider []SpamDomainProvider
+
+// GetSpamDomains returns a combined list of referrer spam domains from all providers.
+func (aggregateProvider AggregateProvider) GetSpamDomains() ([]string, error) {
+
+	var domains []string
+
+	for _, provider := range aggregateProvider {
+		repositoryDomains, err := provider.GetSpamDomains()
+		if err != nil {
+			return nil, err
+		}
+
+		domains = append(domains, repositoryDomains...)
+	}
+
+	sort.Strings(domains)
+	domains = unique(domains)
+
+	return domains, nil
+}
+
+// The remoteSpamDomains provides a static list
 // of referrer spam domain names from a remote URL.
-type staticSpamDomains struct {
+type remoteSpamDomains struct {
 	domainListURL string
 }
 
-// GetSpamDomains returns a list of referrer spam domain names for a
-// remote text file that contains domain names that are considered spam.
-func (provider *staticSpamDomains) GetSpamDomains() ([]string, error) {
+// GetSpamDomains returns a list of referrer spam domain names from a remote text file.
+func (provider *remoteSpamDomains) GetSpamDomains() ([]string, error) {
 
 	// request the domain names from the remote source
 	response, requestError := http.Get(provider.domainListURL)
@@ -51,4 +89,15 @@ func (provider *staticSpamDomains) GetSpamDomains() ([]string, error) {
 	}
 
 	return domains, nil
+}
+
+// The localSpamDomains provides a static list
+// of referrer spam domain names from a local file.
+type localSpamDomains struct {
+	filePath string
+}
+
+// GetSpamDomains returns a list of referrer spam domain names from a local file.
+func (provider *localSpamDomains) GetSpamDomains() ([]string, error) {
+	return nil, nil
 }