+++ spamcontrol/domainrepository.go
@@ -4,40 +4,89 @@ import (
 	"bufio"
 	"fmt"
 	"os"
+	"path/filepath"
 	"sort"
 	"strings"
+
+	"github.com/andreaskoch/ga-spam-control/common/fsutil"
 )
 
-// The SpamDomainRepository interface provides functions for
-// retrieving and storing spam domain names.
-type SpamDomainRepository interface {
-
-	// GetSpamDomains returns a list of referrer spam domains.
-	GetSpamDomains() ([]string, error)
-
-	// UpdateSpamDomains stores the latest list of spam domains.
-	// Returns the additions, removals or an error if the update failed.
-	UpdateSpamDomains() (unchanged, added, removed []string, err error)
-}
-
-// NewSpamDomainRepository creates a new FilesystemSpamDomainRepository instance.
-func NewSpamDomainRepository(filePath string, domainProviderFactory *SpamDomainProviderFactory) SpamDomainRepository {
-	return &FilesystemSpamDomainRepository{
-		filePath:              filePath,
-		domainProviderFactory: domainProviderFactory,
+// NewCommunitySpamDomainRepository creates a new CommunitySpamDomainRepository instance.
+func NewCommunitySpamDomainRepository(filePath string, providers []SpamDomainProvider) *CommunitySpamDomainRepository {
+	return &CommunitySpamDomainRepository{
+		LocalSpamDomainRepository{
+			filePath:  filePath,
+			providers: providers,
+		},
 	}
 }
 
-// A FilesystemSpamDomainRepository provides functions
-// for retrieving and storing spam domain names on disc.
-type FilesystemSpamDomainRepository struct {
-	filePath string
+// NewPrivateSpamDomainRepository creates a new PrivateSpamDomainRepository instance.
+func NewPrivateSpamDomainRepository(filePath string, provider SpamDomainProvider) *PrivateSpamDomainRepository {
+	return &PrivateSpamDomainRepository{
+		LocalSpamDomainRepository{
+			filePath:  filePath,
+			providers: []SpamDomainProvider{provider},
+		},
+	}
+}
 
-	domainProviderFactory *SpamDomainProviderFactory
+// A CommunitySpamDomainRepository stores the referrer spam domains names
+// that are provided by the community locally and provides functions for
+// keeping this file up-to-date.
+type CommunitySpamDomainRepository struct {
+	LocalSpamDomainRepository
 }
 
 // GetSpamDomains returns a list of referrer spam domains from disc.
-func (repository *FilesystemSpamDomainRepository) GetSpamDomains() ([]string, error) {
+func (repository *CommunitySpamDomainRepository) GetSpamDomains() ([]string, error) {
+	return repository.LocalSpamDomainRepository.GetSpamDomains()
+}
+
+// UpdateSpamDomains stores the latest list of spam domains to disc.
+func (repository *CommunitySpamDomainRepository) UpdateSpamDomains() (unchanged, added, removed []string, err error) {
+	return repository.LocalSpamDomainRepository.UpdateSpamDomains()
+}
+
+// A PrivateSpamDomainRepository provides functions for saving a personal
+// list of referrer spam domain names locally.
+type PrivateSpamDomainRepository struct {
+	LocalSpamDomainRepository
+}
+
+// GetSpamDomains returns a list of referrer spam domains from disc.
+func (repository *PrivateSpamDomainRepository) GetSpamDomains() ([]string, error) {
+	return repository.LocalSpamDomainRepository.GetSpamDomains()
+}
+
+// AddDomains add the given spam domain names to the repository.
+func (repository *PrivateSpamDomainRepository) AddDomains(newDomains []string) error {
+	existingDomains, existingDomainsError := repository.LocalSpamDomainRepository.GetSpamDomains()
+	if existingDomainsError != nil {
+		return existingDomainsError
+	}
+
+	mergedDomainNames := append(existingDomains, newDomains...)
+	sort.Strings(mergedDomainNames)
+	mergedDomainNames = unique(mergedDomainNames)
+
+	writeError := repository.LocalSpamDomainRepository.Write(mergedDomainNames)
+	if writeError != nil {
+		return writeError
+	}
+
+	return nil
+}
+
+// A LocalSpamDomainRepository provides functions
+// for retrieving and storing spam domain names on disc.
+type LocalSpamDomainRepository struct {
+	filePath  string
+	providers []SpamDomainProvider
+}
+
+// GetSpamDomains returns a list of referrer spam domains from disc.
+func (repository *LocalSpamDomainRepository) GetSpamDomains() ([]string, error) {
 	domains, err := getDomainsFromFile(repository.filePath)
 	if err != nil {
 
@@ -53,7 +102,7 @@ func (repository *FilesystemSpamDomainRepository) GetSpamDomains() ([]string, er
 }
 
 // UpdateSpamDomains stores the latest list of spam domains to disc.
-func (repository *FilesystemSpamDomainRepository) UpdateSpamDomains() (unchanged, added, removed []string, err error) {
+func (repository *LocalSpamDomainRepository) UpdateSpamDomains() (unchanged, added, removed []string, err error) {
 
 	// get the existing domains for comparision
 	existingDomains, _ := getDomainsFromFile(repository.filePath)
@@ -65,7 +114,7 @@ func (repository *FilesystemSpamDomainRepository) UpdateSpamDomains() (unchanged
 	}
 
 	// write the new list to disc
-	writeError := writeDomainsToFile(repository.filePath, newDomains)
+	writeError := repository.Write(newDomains)
 	if writeError != nil {
 		return nil, nil, nil, writeError
 	}
@@ -74,14 +123,20 @@ func (repository *FilesystemSpamDomainRepository) UpdateSpamDomains() (unchanged
 	return
 }
 
-func (repository *FilesystemSpamDomainRepository) getLatestSpamDomains() ([]string, error) {
-	providers, err := repository.domainProviderFactory.GetSpamDomainProviders()
-	if err != nil {
-		return nil, err
+// Write stores the given domain names in the repository.
+func (repository *LocalSpamDomainRepository) Write(domains []string) error {
+	writeError := writeDomainsToFile(repository.filePath, domains)
+	if writeError != nil {
+		return writeError
 	}
 
+	return nil
+}
+
+func (repository *LocalSpamDomainRepository) getLatestSpamDomains() ([]string, error) {
+
 	var fullDomainList []string
-	for _, provider := range providers {
+	for _, provider := range repository.providers {
 		domains, err := provider.GetSpamDomains()
 		if err != nil {
 			return nil, err
@@ -131,6 +186,15 @@ func diff(slice1, slice2 []string) (unchanged, added, removed []string) {
 // writeDomainsToFile writes a list of domain names to the specified file.
 // Returns an error if the write failed.
 func writeDomainsToFile(filePath string, domains []string) error {
+
+	// make sure the folder exists
+	folder := filepath.Dir(filePath)
+	if !fsutil.PathExists(folder) {
+		if err := os.MkdirAll(folder, 0700); err != nil {
+			return err
+		}
+	}
+
 	file, err := os.OpenFile(filePath, os.O_CREATE|os.O_TRUNC|os.O_RDWR, 0660)
 	if err != nil {
 		return err
@@ -142,7 +206,7 @@ func writeDomainsToFile(filePath string, domains []string) error {
 	defer writer.Flush()
 
 	for _, domain := range domains {
-		fmt.Fprintf(writer, "%s\n", domain)
+		fmt.Fprintf(writer, "%s"+NewLineSequence, domain)
 	}
 
 	return nil